+++ packages/wonder-blocks-tokens/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-tokens
 
+## 4.0.0
+
+### Major Changes
+
+-   0de25cd8: - Reworked semanticColor actions to use a new structure. Every `action` now includes `default`, `hover` and `press` states, and each state includes `border`, `background` and `foreground` tokens.
+
+    -   Renamed `primary` to `progressive`.
+
+    -   Added more categories to actions: `filled` and `outline`.
+
 ## 3.0.1
 
 ### Patch Changes